# frontend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kulturbytes Admin API",
    "description": "Internal reporting API. Production auth is unavailable in milestone 1.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/record-marks": {
      "get": {
        "tags": [
          "Record marks"
        ],
        "summary": "List Marks",
        "operationId": "list_marks_api_v1_record_marks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "organization",
                    "venue",
                    "space",
                    "event",
                    "event_date",
                    "user",
                    "partner_request",
                    "team_membership",
                    "image"
                  ],
                  "type": "string"
                },
                {
                  "enum": [
                    "event_link",
                    "license",
                    "image_link"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1024
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Key"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "open",
                "in_progress",
                "done",
                "all"
              ],
              "type": "string",
              "default": "active",
              "title": "Status"
            }
          },
          {
            "name": "urgency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "normal",
                    "high",
                    "urgent"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Urgency"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "questionable_content",
                    "low_quality",
                    "incorrect",
                    "incomplete",
                    "outdated",
                    "duplicate",
                    "spam",
                    "unsuitable",
                    "rights_privacy",
                    "technical",
                    "other"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "urgency",
                "newest"
              ],
              "type": "string",
              "default": "urgency",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkPage"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Record marks"
        ],
        "summary": "Add Mark",
        "operationId": "add_mark_api_v1_record_marks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/record-marks/{mark_id}": {
      "get": {
        "tags": [
          "Record marks"
        ],
        "summary": "Get Mark",
        "operationId": "get_mark_api_v1_record_marks__mark_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mark Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "patch": {
        "tags": [
          "Record marks"
        ],
        "summary": "Change Mark",
        "operationId": "change_mark_api_v1_record_marks__mark_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mark_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Mark Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDetail"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "Mark": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "organization",
                  "venue",
                  "space",
                  "event",
                  "event_date",
                  "user",
                  "partner_request",
                  "team_membership",
                  "image"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "event_link",
                  "license",
                  "image_link"
                ]
              }
            ],
            "title": "Entity Type"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "entity_name": {
            "type": "string",
            "title": "Entity Name"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "questionable_content",
                "low_quality",
                "incorrect",
                "incomplete",
                "outdated",
                "duplicate",
                "spam",
                "unsuitable",
                "rights_privacy",
                "technical",
                "other"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "normal",
              "high",
              "urgent"
            ],
            "title": "Urgency"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "done"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "entity_key",
          "entity_name",
          "reasons",
          "reason_detail",
          "urgency",
          "status",
          "version",
          "created_at",
          "created_by",
          "updated_at",
          "completed_at",
          "completed_by"
        ],
        "title": "Mark"
      },
      "MarkCreate": {
        "properties": {
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "questionable_content",
                "low_quality",
                "incorrect",
                "incomplete",
                "outdated",
                "duplicate",
                "spam",
                "unsuitable",
                "rights_privacy",
                "technical",
                "other"
              ]
            },
            "type": "array",
            "maxItems": 11,
            "minItems": 1,
            "title": "Reasons"
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "normal",
              "high",
              "urgent"
            ],
            "title": "Urgency",
            "default": "normal"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "organization",
                  "venue",
                  "space",
                  "event",
                  "event_date",
                  "user",
                  "partner_request",
                  "team_membership",
                  "image"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "event_link",
                  "license",
                  "image_link"
                ]
              }
            ],
            "title": "Entity Type"
          },
          "entity_key": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Entity Key"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reasons",
          "entity_type",
          "entity_key"
        ],
        "title": "MarkCreate"
      },
      "MarkDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entity_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "organization",
                  "venue",
                  "space",
                  "event",
                  "event_date",
                  "user",
                  "partner_request",
                  "team_membership",
                  "image"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "event_link",
                  "license",
                  "image_link"
                ]
              }
            ],
            "title": "Entity Type"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "entity_name": {
            "type": "string",
            "title": "Entity Name"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "questionable_content",
                "low_quality",
                "incorrect",
                "incomplete",
                "outdated",
                "duplicate",
                "spam",
                "unsuitable",
                "rights_privacy",
                "technical",
                "other"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "normal",
              "high",
              "urgent"
            ],
            "title": "Urgency"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "done"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed By"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/MarkEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "entity_key",
          "entity_name",
          "reasons",
          "reason_detail",
          "urgency",
          "status",
          "version",
          "created_at",
          "created_by",
          "updated_at",
          "completed_at",
          "completed_by",
          "events"
        ],
        "title": "MarkDetail"
      },
      "MarkEvent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "kind": {
            "type": "string",
            "enum": [
              "created",
              "updated",
              "completed",
              "reopened"
            ],
            "title": "Kind"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "done"
            ],
            "title": "Status"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "questionable_content",
                "low_quality",
                "incorrect",
                "incomplete",
                "outdated",
                "duplicate",
                "spam",
                "unsuitable",
                "rights_privacy",
                "technical",
                "other"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "normal",
              "high",
              "urgent"
            ],
            "title": "Urgency"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "kind",
          "author",
          "created_at",
          "note",
          "status",
          "reasons",
          "reason_detail",
          "urgency"
        ],
        "title": "MarkEvent"
      },
      "MarkPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Mark"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "MarkPage"
      },
      "MarkUpdate": {
        "properties": {
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "questionable_content",
                "low_quality",
                "incorrect",
                "incomplete",
                "outdated",
                "duplicate",
                "spam",
                "unsuitable",
                "rights_privacy",
                "technical",
                "other"
              ]
            },
            "type": "array",
            "maxItems": 11,
            "minItems": 1,
            "title": "Reasons"
          },
          "reason_detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Detail"
          },
          "urgency": {
            "type": "string",
            "enum": [
              "normal",
              "high",
              "urgent"
            ],
            "title": "Urgency",
            "default": "normal"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "done"
            ],
            "title": "Status"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reasons",
          "version",
          "status"
        ],
        "title": "MarkUpdate"
      },
...
      "Pagination": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "pages"
        ],
        "title": "Pagination"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}